WI #2344 Make PICTURE character sequence public
